Refactored many of the CwmsDataManager methods into individual Dao classes.
