+++ cwms_radar_api/src/main/java/cwms/radar/api/OfficeController.java
@@ -1,41 +1,44 @@
 package cwms.radar.api;
 
-import io.javalin.apibuilder.CrudHandler;
-import io.javalin.core.util.Header;
-import io.javalin.http.Context;
-import io.javalin.http.NotFoundResponse;
-import io.javalin.plugin.openapi.annotations.*;
-
 import java.io.IOException;
-import java.sql.SQLException;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-
 import javax.servlet.http.HttpServletResponse;
 
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
-import static com.codahale.metrics.MetricRegistry.*;
 import com.codahale.metrics.Timer;
-
-import cwms.radar.data.CwmsDataManager;
+import cwms.radar.data.dao.OfficeDao;
 import cwms.radar.data.dto.Office;
+import cwms.radar.formatters.ContentType;
 import cwms.radar.formatters.Formats;
 import cwms.radar.formatters.FormattingException;
 import cwms.radar.formatters.OfficeFormatV1;
 import cwms.radar.formatters.csv.CsvV1Office;
 import cwms.radar.formatters.tab.TabV1Office;
-import cwms.radar.formatters.ContentType;
+import io.javalin.apibuilder.CrudHandler;
+import io.javalin.core.util.Header;
+import io.javalin.http.Context;
+import io.javalin.http.NotFoundResponse;
+import io.javalin.plugin.openapi.annotations.OpenApi;
+import io.javalin.plugin.openapi.annotations.OpenApiContent;
+import io.javalin.plugin.openapi.annotations.OpenApiParam;
+import io.javalin.plugin.openapi.annotations.OpenApiResponse;
+import org.jooq.DSLContext;
+
+import static com.codahale.metrics.MetricRegistry.name;
+import static cwms.radar.data.dao.JooqDao.getDslContext;
+
 /**
  *
  */
 public class OfficeController implements CrudHandler {
     private static final Logger logger = Logger.getLogger(OfficeController.class.getName());
-    private final MetricRegistry metrics;// = new MetricRegistry();
-    private final Meter getAllRequests;// = metrics.meter(OfficeController.class.getName()+"."+"getAll.count");
-    private final Timer getAllRequestsTime;// =metrics.timer(OfficeController.class.getName()+"."+"getAll.time");
+    private final MetricRegistry metrics;
+    private final Meter getAllRequests;
+    private final Timer getAllRequestsTime;
     private final Meter getOneRequest;
     private final Timer getOneRequestTime;
     private final Histogram requestResultSize;
@@ -85,23 +88,20 @@ public class OfficeController implements CrudHandler {
         getAllRequests.mark();
         
         try (
-                final Timer.Context time_context  = getAllRequestsTime.time();
-                CwmsDataManager cdm = new CwmsDataManager(ctx);
-            ) {                                            
-                List<Office> offices = cdm.getOffices();
-                String format_parm = ctx.queryParam("format","");
-                String format_header = ctx.header(Header.ACCEPT);
-                ContentType contentType = Formats.parseHeaderAndQueryParm(format_header, format_parm);                
+                final Timer.Context timeContext  = getAllRequestsTime.time();
+                DSLContext dsl = getDslContext(ctx))
+        {
+                OfficeDao dao = new OfficeDao(dsl);
+                List<Office> offices = dao.getOffices();
+                String formatParm = ctx.queryParam("format","");
+                String formatHeader = ctx.header(Header.ACCEPT);
+                ContentType contentType = Formats.parseHeaderAndQueryParm(formatHeader, formatParm);
                 
                 String result = Formats.format(contentType,offices);
                 
                 ctx.result(result).contentType(contentType.toString());
                 requestResultSize.update(result.length());
-                
-        } catch (SQLException ex) {
-            logger.log(Level.SEVERE, null, ex);
-            ctx.status(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-            ctx.result("Failed to process request");
+
         } catch ( FormattingException fe ){
             logger.log(Level.SEVERE,"failed to format data",fe);
             if( fe.getCause() instanceof IOException ){
@@ -137,27 +137,24 @@ public class OfficeController implements CrudHandler {
         tags = {"Offices"}
     )
     @Override
-    public void getOne(Context ctx, String office_id) {
+    public void getOne(Context ctx, String officeId) {
         getOneRequest.mark();
         try(
-            final Timer.Context time_context = getOneRequestTime.time();
-            CwmsDataManager cdm = new CwmsDataManager(ctx);
-        ) {            
-            Office office = cdm.getOfficeById(office_id);
+            final Timer.Context timeContext = getOneRequestTime.time();
+            DSLContext dsl = getDslContext(ctx))
+        {
+            OfficeDao dao = new OfficeDao(dsl);
+            Office office = dao.getOfficeById(officeId);
             if( office == null ){
-                throw new NotFoundResponse("Office id: " + office_id + " does not exist");
+                throw new NotFoundResponse("Office id: " + officeId + " does not exist");
             }
-            String format_parm = ctx.queryParam("format","");
-            String format_header = ctx.header(Header.ACCEPT);
-            ContentType contentType = Formats.parseHeaderAndQueryParm(format_header, format_parm);                                                    
+            String formatParm = ctx.queryParam("format","");
+            String formatHeader = ctx.header(Header.ACCEPT);
+            ContentType contentType = Formats.parseHeaderAndQueryParm(formatHeader, formatParm);
             String result = Formats.format(contentType,office);
             ctx.result(result).contentType(contentType.toString());
-            //ctx.result(result).contentType(formatter.getContentType());
+
             requestResultSize.update(result.length());
-        } catch( SQLException ex ){
-            Logger.getLogger(LocationController.class.getName()).log(Level.SEVERE, null, ex);
-            ctx.status(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-            ctx.result("Failed to process request");
         } catch( FormattingException fe ){
             logger.log(Level.SEVERE,"failed to format data",fe);
             if( fe.getCause() instanceof IOException ){
@@ -178,13 +175,13 @@ public class OfficeController implements CrudHandler {
 
     @OpenApi(ignore = true)
     @Override
-    public void update(Context ctx, String location_code) {
+    public void update(Context ctx, String officeId) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @OpenApi(ignore = true)
     @Override
-    public void delete(Context ctx, String location_code) {
+    public void delete(Context ctx, String officeId) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
